requirements.txt
django==5.0.9
djangorestframework==3.15.2
requests==2.32.3
numpy==2.2.0
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
tzdata==2024.2
pymysql==1.1.1
et-xmlfile==2.0.0
openpyxl==3.1.5
django-cors-headers==4.6.0


mysqlclient==2.2.4
SQLAlchemy==1.4.54
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
redis==5.0.8
drf-nested-routers==0.94.1
ddtrace<2.10
urllib3==1.26.19
tenacity==9.0.0